ON-15122: Add unit tests for upgrade logic
This change adds a new set of unit tests to test the upgrade logic. In
addition to the existing tests that use EnvTest to provide the API
server, I have added gomock as new dependency for this project.

Without any mocking of the controller-runtime client we would be unable
to properly test the controller as it has to manage labels on nodes and
lifetimes of pods, and using EnvTest alone we are unable to test this
since it doesn't manage Nodes or other controllers.

The mocks for the controller-runtime client interface were automatically
generated and added as a new file in mocks/client/mock_client.go.

Unit tests that use gomock were added for functions with low code
coverage, and are not yet exhaustive of all code paths.